# openapi/nova-browser-api.json
{
  "components": {
    "schemas": {
...
      "DurationRequest": {
        "properties": {
          "duration_days": {
            "items": {
              "type": "number"
            },
            "title": "Duration Days",
            "type": "array"
          }
        },
        "required": [
          "duration_days"
        ],
        "title": "DurationRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FloodDepthRequest": {
        "properties": {
          "ground_elevation": {
            "items": {
              "type": "number"
            },
            "title": "Ground Elevation",
            "type": "array"
          },
          "water_surface_elevation": {
            "items": {
              "type": "number"
            },
            "title": "Water Surface Elevation",
            "type": "array"
          }
        },
        "required": [
          "water_surface_elevation",
          "ground_elevation"
        ],
        "title": "FloodDepthRequest",
        "type": "object"
      },
...
      "ScreeningAtlasRequest": {
        "properties": {
          "aoi_id": {
            "format": "uuid",
            "title": "Aoi Id",
            "type": "string"
          },
          "baseline_end": {
            "format": "date",
            "title": "Baseline End",
            "type": "string"
          },
          "baseline_start": {
            "format": "date",
            "title": "Baseline Start",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "target_end": {
            "format": "date",
            "title": "Target End",
            "type": "string"
          },
          "target_start": {
            "format": "date",
            "title": "Target Start",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "aoi_id",
          "target_start",
          "target_end",
          "baseline_start",
          "baseline_end"
        ],
        "title": "ScreeningAtlasRequest",
        "type": "object"
      },
...
      "TaskTriggerResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskTriggerResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NOVA-GeoRisk Browser API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingestion/screening-atlas": {
      "post": {
        "description": "Starts a flood screening atlas job for the given project's AOI.\nWorks for any project/AOI \u2014 nothing here is specific to any one\nlocation; the AOI's own stored geometry is what gets analyzed.",
        "operationId": "trigger_screening_atlas_api_v1_ingestion_screening_atlas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreeningAtlasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTriggerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Trigger Screening Atlas",
        "tags": [
          "Ingestion"
        ]
      }
    },
    "/api/v1/maps/atlas/{project_id}/{aoi_id}": {
      "get": {
        "description": "Generic screening-atlas manifest for any project/AOI \u2014 reads\nwhatever `generate_flood_screening_atlas()` (triggered via\n`POST /ingestion/screening-atlas`) wrote to\n`{OUTPUT_STORAGE_DIR}/{project_id}/{aoi_id}/metadata.json`.\nWorks for any AOI, not one specific location.",
        "operationId": "get_flood_screening_atlas_api_v1_maps_atlas__project_id___aoi_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "aoi_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Aoi Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtlasManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Flood Screening Atlas",
        "tags": [
          "Flood Maps"
        ]
      }
    },
    "/api/v1/maps/atlas/{project_id}/{aoi_id}/products/{product_key}": {
      "get": {
        "operationId": "get_legacy_atlas_product_api_v1_maps_atlas__project_id___aoi_id__products__product_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "aoi_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Aoi Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "product_key",
            "required": true,
            "schema": {
              "title": "Product Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Legacy Atlas Product",
        "tags": [
          "Flood Maps"
        ]
      }
    },
    "/api/v1/maps/depth": {
      "post": {
        "description": "Depth = WSE - Ground \u2014 Doc 2 Map 2.",
        "operationId": "flood_depth_api_v1_maps_depth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloodDepthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifiedValuesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Flood Depth",
        "tags": [
          "Flood Maps"
        ]
      }
    },
    "/api/v1/maps/duration": {
      "post": {
        "description": "Flood duration classification \u2014 Doc 2 Map 6.",
        "operationId": "flood_duration_api_v1_maps_duration_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DurationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifiedValuesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Flood Duration",
        "tags": [
          "Flood Maps"
        ]
      }
    },
    "/api/v1/maps/exposure": {
      "post": {
        "description": "Population/asset density classification \u2014 Doc 2 Map 8.",
        "operationId": "flood_exposure_density_api_v1_maps_exposure_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExposureDensityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifiedValuesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Flood Exposure Density",
        "tags": [
          "Flood Maps"
        ]
      }
    },
    "/api/v1/maps/extent": {
      "post": {
        "description": "SAR backscatter-ratio flood detection \u2014 Doc 2 Map 1.",
        "operationId": "flood_extent_api_v1_maps_extent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloodExtentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloodExtentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Flood Extent",
        "tags": [
          "Flood Maps"
        ]
      }
    },
...
  }
}